refactor: (viva-ms) convert createCase to typescript
